test(api): merge integ and unit test on ChorusAdapter
